# usf/src/main/java/uol/compass/project/usf/model/dto/request/UserRequestDTO.java
package uol.compass.project.usf.model.dto.request;

import javax.validation.constraints.NotBlank;

import lombok.AllArgsConstructor;
import lombok.Builder;
import lombok.Data;
import lombok.NoArgsConstructor;

@Data
@Builder
@NoArgsConstructor
@AllArgsConstructor
public class UserRequestDTO {

    @NotBlank
    private String name;

    @NotBlank
    private String login;

    @NotBlank
    private String password;

}