# usf/src/main/java/uol/compass/project/usf/model/dto/response/UserResponseDTO.java
package uol.compass.project.usf.model.dto.response;

import java.util.List;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;
import uol.compass.project.usf.model.entities.Role;

@Data
@AllArgsConstructor
@NoArgsConstructor
public class UserResponseDTO {

    private Long id;

    private String name;

    private String login;

    private String password;

    private List<Role> roles;

}