# usf/src/main/java/uol/compass/project/usf/model/entities/DoctorEntity.java
package uol.compass.project.usf.model.entities;

import lombok.AllArgsConstructor;
import lombok.Data;
import lombok.NoArgsConstructor;

import javax.persistence.*;

import com.fasterxml.jackson.annotation.JsonIgnore;

@Data
@AllArgsConstructor
@NoArgsConstructor
@Entity(name = "doctor")
@Table(name = "doctor")
public class DoctorEntity {

    @Id
    @GeneratedValue(strategy = GenerationType.IDENTITY)
    private Long id;

    private String name;

    private String specialization;

    @ManyToOne
    @JoinColumn(name = "team_id")
    @JsonIgnore
    private TeamEntity team;

}