# usf/src/main/resources/swagger.yml
openapi: 3.0.3
info:
  title: Swagger USF API - OpenAPI 0.0.1
  description: |-
    A USF (Unidade de Saúde da Família) nada mais é que uma unidade pública de saúde destinada
    a realizar atenção contínua nas especialidades básicas, com uma equipe multiprofissional
    habiitada para desenvolver as atividades de promoção, proteção e recuperação, características
    do nível primário de atenção. Após conversa com funcionários dessas unidades notamos que há uma
    complexidade na gestão desse ambiente, pois precisam fazer a comunicação com outras USFs, seja para
    realizar pedidos de recursos ou para gestão de profissionais, de maneira manual através de pranchetas 
    e documentos por escrito. Fazendo o sistema desses ambientes vitais para o cuidado das comunidades brasileiras 
    ficam propensos a erros e lentidão no atendimento.

    Objetivando desenvolvimento para esse serviço, desenvolvemos uma API para automatizar esses processos defasados.


    Alguns links que podem ser interessantes:
    - [Repositório da USF API](https://github.com/solanokruger/usf-project)
    - [Toda documentação do projeto](https://docs.google.com/document/d/1Rgm3JiFKQBkO_tDVDDw1Mr-FGm9hP7MV-BGx2f4m3HE/edit?usp=sharing)
  termsOfService: http://swagger.io/terms/
  contact:
    email: luis.gustavo@compasso.com.br
  version: 0.0.1
servers:
  - url: http://localhost:8080
    description: API local
...
paths:
...
  /doctor{idDoctor}/team/{idTeam}:
    post:
      security:
        - bearerAuth: []
      tags:
        - Doctor
      summary: Vincula um Doutor a um Time
      description: Vincula um Doutor a um Time
      operationId: linkDoctorTeam
      parameters:
        - name: idDoctor
          in: path
          description: ID do Doutor a ser vinculado.
          required: true
          schema:
            type: integer
            format: int64
        - name: idTeam
          in: path
          description: ID do Time que o Doutor irá ser vinculado.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorResponseDTO"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: DOCTOR OR TEAM NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Doctor
      summary: Remove o Doutor de um Time
      description: Remove um Doutor de um time específico.
      operationId: deleteDoctorTeam
      parameters:
        - name: idDoctor
          in: path
          description: ID do Doutor a ser removido.
          required: true
          schema:
            type: integer
            format: int64
        - name: idTeam
          in: path
          description: ID do time que o Doutor irá ser removido.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/DoctorResponseDTO"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: DOCTOR OR TEAM NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"
  /team:
    post:
      security:
        - bearerAuth: []
      tags:
        - Team
      summary: Adiciona um novo time.
      description: Adiciona um novo time.
      operationId: addTeam
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequestDTO"
      responses:
        "200":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"

    get:
      security:
        - bearerAuth: []
      tags:
        - Team
      summary: Retorna todos os times
      description: Retorna uma lista com todos os times
      operationId: getTeam
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/TeamResponseDTO"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
  /team/{idTeam}:
    get:
      security:
        - bearerAuth: []
      tags:
        - Team
      summary: Procura um time pelo ID
      description: Retorna apenas um time
      operationId: getTeamById
      parameters:
        - name: idTeam
          in: path
          description: ID do time para Requisição.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: TEAM NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"

    put:
      security:
        - bearerAuth: []
      tags:
        - Team
      summary: Atualiza um time
      description: Atualiza um time existente pelo ID
      operationId: updateTeam
      parameters:
        - name: idTeam
          in: path
          description: ID do time para Requisição.
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        required: true
        description: Atualiza o time
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/TeamRequestDTO"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/TeamResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: TEAM NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"

    delete:
      security:
        - bearerAuth: []
      tags:
        - Team
      summary: Remove um time
      description: Remove um time
      operationId: deleteTeam
      parameters:
        - name: idTeam
          in: path
          description: ID do time para deletar.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: TEAM NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"
  /team/{idTeam}/doctor:
    get:
      security:
        - bearerAuth: []
      tags:
        - Team
      summary: Retorna todos os doutores em um time
      description: Retorna uma lista com todos os doutores de um time
      operationId: getDoctorsByTeam
      parameters:
        - name: idTeam
          in: path
          description: ID do time para Requisição.
          required: true
          schema:
            type: integer
            format: int64
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/DoctorResponseDTO"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: TEAM NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"
...
  /user/create:
    post:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: Cria usuário
      description: Usuários só podem ser criados pelo administrador
      operationId: createUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/UserRequestDTO"
      responses:
        "200":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
  /user/update/{id}:
    post:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: Atualiza usuário
      description: Utualiza dados do usuário por id
      operationId: updateUser
      parameters:
        - name: id
          in: path
          description: ID do usuário para atualizar
          required: true
          schema:
            type: integer
            format: int64
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: "#/components/schemas/UserRequestUpdateDTO"
      responses:
        "200":
          description: OK
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
        "404":
          description: USER NOT FOUND
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/NotFoundException"
  /user/role:
    post:
      security:
        - bearerAuth: []
      tags:
        - User
      summary: Define cargos no sistema ao usuário
      description: Define cargos no sistema ao usuário através do id
      operationId: setRoleUser
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/CreateUserRoleDTO"
      responses:
        "200":
          description: CREATED
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/UserResponseDTO"
        "400":
          description: BAD REQUEST
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/MethodArgumentNotValidException"
        "401":
          $ref: "#/components/schemas/UnauthorizedError"
...
components:
  schemas:
...
    TeamRequestDTO:
      type: object
      properties:
        color:
          type: string
      example:
        color: azul
...
    UserRequestDTO:
      type: object
      properties:
        name:
          type: string
        login:
          type: string
        password:
          type: string
      example:
        name: Gustavo
        login: admin
        password: 12345
    UserRequestUpdateDTO:
      type: object
      properties:
        oldPassword:
          type: string
        name:
          type: string
        newPassword:
          type: string
      example:
        oldPassword: 12345
        name: Miguel
        password: 54321
    CreateUserRoleDTO:
      type: object
      properties:
        idUser:
          type: integer
          format: int64
        idRoles:
          type: array
          items:
            type: integer
            format: int64
      example:
        idUser: 1
        idRoles: [1, 2, 3]
    TeamEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        color:
          type: string
      example:
        id: 1
        color: azul
    DoctorResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        specialization:
          type: string
        team:
          $ref: "#/components/schemas/TeamEntity"
      example:
        id: 1
        name: Társila
        specialization: Oftalmologista
...
    UsfEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        address:
          type: string
      example:
        id: 1
        name: Unidade de Saúde da Família da Avenida 2 de Julho
        address: Avenida 2 de Julho, número 15
...
    DoctorEntity:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        specilization:
          type: string
      example:
        id: 1
        name: Marcos
        specialization: Cardiologista
    TeamResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        color:
          type: string
        currentUsf:
          $ref: "#/components/schemas/UsfEntity"
        doctors:
          type: array
          items:
            $ref: "#/components/schemas/DoctorEntity"
      example:
        id: 1
        color: azul
...
    Role:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
    UserResponseDTO:
      type: object
      properties:
        id:
          type: integer
          format: int64
        name:
          type: string
        login:
          type: string
        password:
          type: string
        roles:
          type: array
          items:
            $ref: "#/components/schemas/Role"
    MethodArgumentNotValidException:
      type: object
      properties:
        details:
          type: string
        errorCode:
          type: string
        httpStatus:
          type: string
    NotFoundException:
      type: object
      properties:
        details:
          type: string
        errorCode:
          type: string
        httpStatus:
          type: string
...
    UnauthorizedError:
      description: Access token is missing or invalid
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT